- Updated indentation - Updated proto files to remove akkaserverless annotations - Updated config.yml for CI
