+++ samples/java-valueentity-counter-spring-client/src/main/java/com/example/client/spring/rest/CounterClientApplication.java
@@ -1,0 +1,11 @@
+package com.example.client.spring.rest;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CounterClientApplication {
+  public static void main(String[] args) {
+    SpringApplication.run(CounterClientApplication.class, args);
+  }
+}
