+++ samples/java-valueentity-counter-spring-client/src/main/proto/com/example/actions/counter_states_sub.proto
@@ -1,0 +1,31 @@
+// Copyright 2021 Lightbend Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// tag::state-subscription[]
+syntax = "proto3";
+package com.example.actions;
+
+import "com/example/domain/counter_domain.proto"; // <1>
+import "google/protobuf/empty.proto";
+
+option java_outer_classname = "StateSubscriptionApi";
+
+service CounterStateSubscription {
+
+  
+  rpc OnUpdateState (com.example.domain.CounterState) returns (google.protobuf.Empty) {
+  } 
+
+}
+// end::state-subscription[]
