+++ samples/java-valueentity-counter-spring-client/src/main/proto/com/example/counter_api.proto
@@ -1,0 +1,38 @@
+// This is the public API offered by your entity.
+syntax = "proto3";
+
+package com.example; // <1>
+
+import "google/api/annotations.proto";
+import "google/protobuf/empty.proto";
+
+option java_outer_classname = "CounterApi"; // <3>
+
+message IncreaseValue { // <4>
+  string counter_id = 1; // <5>
+  int32 value = 2;
+}
+
+message DecreaseValue {
+  string counter_id = 1;
+  int32 value = 2;
+}
+
+message ResetValue {
+  string counter_id = 1;
+}
+
+message GetCounter {
+  string counter_id = 1;
+}
+
+message CurrentCounter { // <6>
+  int32 value = 1;
+}
+
+service CounterService { // <7>
+  rpc Increase (IncreaseValue) returns (google.protobuf.Empty);
+  rpc Decrease (DecreaseValue) returns (google.protobuf.Empty);
+  rpc Reset (ResetValue) returns (google.protobuf.Empty);
+  rpc GetCurrentCounter (GetCounter) returns (CurrentCounter);
+}
