+++ samples/java-valueentity-counter-spring-client/src/main/proto/com/example/domain/counter_domain.proto
