+++ samples/java-valueentity-counter-spring-client/src/main/resources/application.properties
