+++ samples/java-valueentity-counter-spring-client/src/test/resources/mockito-extensions/org.mockito.plugins.MockMaker
